+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @websocks/server
 
+## 0.2.11
+
+### Patch Changes
+
+- final react patch
+
 ## 0.2.10
 
 ### Patch Changes